msm: add userspace undervolting interface
First of all, this work is havily based on this commit; franciscofranco/mako@48e0074

This commit adds a undervolting interface for the moto g. Normally the kernel
would set the right voltages at boot (during some sort of configuration phase).
Although AVS works as it should, we can still undervolt a little bit more.
A quick dmesg | grep cpr_voltage_plan_init reveals the current voltages for the cpu.
This is set to the lowest possible value (1140000 uV).

Starting with the lowest value for the voltage;
echo 1140 1140 1140 1140 1140 1140 1140 > /sys/devices/system/cpu/cpu0/cpufreq/UV_mV_table

I set the default values to 1140, 1150, 1280 which seem to be the same as before
AVS kicks in. I am sure we can take the AVS-process into account and set the
correct voltages for each frequency and device during boot with this interface.

Conflicts:
	arch/arm/mach-msm/Kconfig
	arch/arm/mach-msm/acpuclock-8226.c

Signed-off-by: Apavayan <info@apavayan.com>